Set minor tick marks on both axis
This commit adds minor tick marks on both the y and x axis to make
reading the graphs a little easier. It was a little tricky to count
exactly how many bugs were in each group with just major tick marks
on the y axis so this adds a minor tick for every 10 bugs. On the x
axis it sets a minor tick for each day, this way it's easy to figure
out each day's data on the graph.